Support workers running on multiple hosts
The basic support for this is already there but the hostname is not
respected in all cases -> this PR fills this gap.

Since each host controls their workers, it's easy to just read/write
the keys with the hostname added.
Due to this change for workers, there's no adaption
`isRunningSchedulerWorker` necessary.

One more word regarding the scheduler:
- this change still supports that only one scheduler gets started!
- but if you try to start many schedulers in parallel when none is there
  yet, they still may start up multiple times, because the redis checks
  together with the PHP code are not atomic.
  My recommendation is: if you care about having a guaranteed single
  scheduler, use a different approach which only starts one.

Regarding compatibility: this change is a BREAKING CHANGEð

In order to apply it to an existing system, jobs/workers have to drained
first and then start up with the new code.

This PR requires with https://github.com/wa0x6e/Cake-Resque/pull/93